drivechain_client: add tooltips to icon buttons
